backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pydantic==2.9.2
python-jose==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
